MAP -> CMB2: move map logic to map model
